+++ scheme.html
@@ -33,7 +33,7 @@
   <body>
     <div id="app" class="overall-layout"></div>
     <script type="module">
-      import App from "./src/App.svelte";
+      import App from "./src/pages/App.svelte";
 
       new App({
         target: document.getElementById("app"),
