+++ src/lib/common/CollapsibleCard.svelte
