+++ src/lib/common/DiscreteLegend.svelte
