+++ src/lib/common/README.md
@@ -1,0 +1,1 @@
+These are generic components that might belong in (or be replaced by) a Svelte component library.
