+++ src/lib/sidebar/About.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import Modal from "./Modal.svelte";
-  import ExternalLink from "./ExternalLink.svelte";
+  import Modal from "../common/Modal.svelte";
+  import ExternalLink from "../common/ExternalLink.svelte";
 
   export let open: boolean;
 </script>
