+++ src/pages/App.svelte
@@ -1,26 +1,26 @@
 <script lang="ts">
   import type { FeatureCollection, Polygon } from "geojson";
   import { onMount } from "svelte";
-  import authoritiesUrl from "../assets/authorities.geojson?url";
-  import type { Schema } from "./types";
+  import authoritiesUrl from "../../assets/authorities.geojson?url";
+  import type { Schema } from "../types";
   import * as Comlink from "comlink";
-  import workerWrapper from "./worker?worker";
-  import { type RouteInfo } from "./worker";
+  import workerWrapper from "../worker?worker";
+  import { type RouteInfo } from "../worker";
 
-  import About from "./lib/About.svelte";
-  import Instructions from "./lib/Instructions.svelte";
-  import Map from "./lib/Map.svelte";
-  import Layout from "./lib/Layout.svelte";
-  import BoundaryLayer from "./lib/BoundaryLayer.svelte";
-  import InterventionList from "./lib/InterventionList.svelte";
-  import EntireScheme from "./lib/EntireScheme.svelte";
-  import ZoomOutMap from "./lib/ZoomOutMap.svelte";
-  import BaselayerSwitcher from "./lib/BaselayerSwitcher.svelte";
-  import Legend from "./lib/Legend.svelte";
+  import About from "../lib/sidebar/About.svelte";
+  import Instructions from "../lib/sidebar/Instructions.svelte";
+  import Map from "../lib/Map.svelte";
+  import Layout from "../lib/common/Layout.svelte";
+  import BoundaryLayer from "../lib/BoundaryLayer.svelte";
+  import InterventionList from "../lib/sidebar/InterventionList.svelte";
+  import EntireScheme from "../lib/sidebar/EntireScheme.svelte";
+  import ZoomOutMap from "../lib/ZoomOutMap.svelte";
+  import BaselayerSwitcher from "../lib/BaselayerSwitcher.svelte";
+  import Legend from "../lib/Legend.svelte";
 
-  import InterventionLayer from "./lib/draw/InterventionLayer.svelte";
-  import HoverLayer from "./lib/draw/HoverLayer.svelte";
-  import Toolbox from "./lib/draw/Toolbox.svelte";
+  import InterventionLayer from "../lib/draw/InterventionLayer.svelte";
+  import HoverLayer from "../lib/draw/HoverLayer.svelte";
+  import Toolbox from "../lib/draw/Toolbox.svelte";
 
   let showAbout = false;
   let showInstructions = false;
